+++ Server/Excel2Json/Data/ApplicationDbContext.cs
@@ -1,4 +1,4 @@
-﻿using Excel2Json.Models;
+﻿using Excel2Json.Domain;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
