+++ Server/Excel2Json/Services/TokenService.cs
@@ -1,4 +1,5 @@
-﻿using Google.Apis.Auth;
+﻿using Excel2Json.Domain;
+using Google.Apis.Auth;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.Extensions.Configuration;
 using Microsoft.IdentityModel.Tokens;
@@ -8,7 +9,7 @@ using System.Security.Claims;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Excel2Json.Auth
+namespace Excel2Json.Services
 {
     public interface ITokenService
     {
