+++ Server/Excel2Json/Startup.cs
@@ -1,5 +1,6 @@
-using Excel2Json.Auth;
 using Excel2Json.Data;
+using Excel2Json.Domain;
+using Excel2Json.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
